Extract ui related things into core ui module from common
