+++ app/src/main/java/com/alefimenko/iuttimetable/feature/RootActivity.kt
@@ -8,7 +8,7 @@ import com.alefimenko.iuttimetable.R
 import com.alefimenko.iuttimetable.core.data.NetworkStatusReceiver
 import com.bluelinelabs.conductor.Conductor
 import com.bluelinelabs.conductor.Router
-import io.github.steelahhh.common.base.BaseActivity
+import com.alefimenko.iuttimetable.base.BaseActivity
 import org.koin.android.ext.android.inject
 
 /*
