+++ app/src/main/java/com/alefimenko/iuttimetable/feature/schedule/schedulepage/SchedulePageController.kt
@@ -6,7 +6,7 @@ import android.view.ViewGroup
 import android.widget.TextView
 import com.alefimenko.iuttimetable.R
 import com.alefimenko.iuttimetable.core.data.DateInteractor
-import io.github.steelahhh.common.base.BaseController
+import com.alefimenko.iuttimetable.base.BaseController
 import org.koin.android.ext.android.inject
 
 /*
