+++ app/src/main/res/drawable/ic_pick_date.xml
@@ -1,7 +1,7 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
     android:width="24dp"
     android:height="24dp"
-    android:tint="#FFFFFF"
+    android:tint="?colorOnBackground"
     android:viewportWidth="24.0"
     android:viewportHeight="24.0">
     <path
