+++ app/src/main/res/drawable/shape_pill_indicator_dark.xml
@@ -9,7 +9,7 @@
             <solid android:color="@android:color/transparent" />
             <stroke
                 android:width="1dp"
-                android:color="@color/backgroundLight" />
+                android:color="@color/backgroundDark" />
             <corners
                 android:bottomLeftRadius="50dp"
                 android:bottomRightRadius="50dp"
