+++ buildSrc/src/main/kotlin/Config.kt
@@ -8,6 +8,7 @@ object ApplicationID {
 object Modules {
     const val app = ":app"
     const val common = ":common"
+    const val coreUi = ":coreui"
     const val local = ":data:local"
 }
