+++ common/src/main/AndroidManifest.xml
@@ -1,1 +1,1 @@
-<manifest package="io.github.steelahhh.common" />
+<manifest package="com.alefimenko.iuttimetable.common" />
