+++ coreui/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="io.github.steelahhh.coreui" />
