+++ coreui/src/main/res/drawable/splash.xml
