+++ coreui/src/main/res/font/manrope_medium.otf
