+++ coreui/src/main/res/font/manrope_regular.otf
