+++ coreui/src/main/res/font/manrope_thin.otf
