+++ coreui/src/main/res/layout/stub_error.xml
