+++ coreui/src/main/res/values-v23/styles.xml
