+++ coreui/src/main/res/values/arrays.xml
