+++ coreui/src/main/res/values/colors.xml
